requirements.txt
anthropic
cohere
fastapi
flask
google-generativeai
groq
httpx
joblib
matplotlib
numpy
pandas
plotly
pydantic
python-dotenv
requests
scikit-learn
torch
tqdm
transformers
uvicorn
werkzeug
